+++ tests/fn/test_fn.py
@@ -7,7 +7,7 @@ from aioway.fate import Fate
 from aioway.fn import (
     FateFn,
     Fn,
-    FnHistory,
+    HistoryGraph,
     MightFn,
     NnFwdFn,
     NnInitFn,
@@ -32,7 +32,7 @@ def _fn_cls():
 
 
 def _input_cls():
-    yield FnHistory
+    yield HistoryGraph
     yield TorchThunk
     yield FateFn
     yield NnFwdFn
